remove useless file, test=develop
